refactor: add transactions, logging, readme and fix bugs

+++ services/src/main/java/com/epam/esm/services/daos/GiftCertificateDAO.java
@@ -13,5 +13,5 @@ public interface GiftCertificateDAO {
 
   int updateGiftCertificate(GiftCertificate giftCertificate);
 
-  int createGiftCertificate(GiftCertificate giftCertificate);
+  Long createGiftCertificate(GiftCertificate giftCertificate);
 }
